refactor: move CsvUpload to @blocks
It imports only @ui, @schemas/csvUpload, and react-dropzone -- it knows nothing
about bank transactions. COMPONENT_MIGRATION_PLAN.md filed it under the
bankTransactions domain because UploadTransactions is its only consumer, but the
layering rule is about what a module knows, not who calls it.

Five components each get their own directory; validateCsvFile.ts nests with
CsvUpload. Names and CSS classes unchanged -- componentName='ValidateCsvTable'
still generates the same Layer__UI__Table-Cell__ValidateCsvTable--* selectors,
which uploadTransactions.scss styles.
