+++ src/components/UploadTransactions/UploadTransactionsUploadCsvStep.tsx
@@ -6,10 +6,10 @@ import { SubmitButton } from '@ui/Button/SubmitButton'
 import { Separator } from '@ui/Separator/Separator'
 import { HStack, Spacer, VStack } from '@ui/Stack/Stack'
 import { Label, P } from '@ui/Typography/Text'
+import { CopyTemplateHeadersButtonGroup } from '@blocks/CsvUpload/CopyTemplateHeadersButtonGroup/CopyTemplateHeadersButtonGroup'
+import { CsvUpload } from '@blocks/CsvUpload/CsvUpload/CsvUpload'
+import { DownloadCsvTemplateButton } from '@blocks/CsvUpload/DownloadCsvTemplateButton/DownloadCsvTemplateButton'
 import { useWizard } from '@blocks/Wizard/Wizard'
-import { CopyTemplateHeadersButtonGroup } from '@components/CsvUpload/CopyTemplateHeadersButtonGroup'
-import { CsvUpload } from '@components/CsvUpload/CsvUpload'
-import { DownloadCsvTemplateButton } from '@components/CsvUpload/DownloadCsvTemplateButton'
 import { allHeaders, templateExampleTransactions, templateHeaders } from '@components/UploadTransactions/template'
 import { type AccountOption } from '@features/customAccounts/CustomAccountComboBox/AccountOption'
 import { CustomAccountComboBox } from '@features/customAccounts/CustomAccountComboBox/CustomAccountComboBox'
