+++ src/components/blocks/CsvUpload/ValidateCsvTable/validateCsvTable.scss
